fix: BodyType 쿼리 시 경계값 처리
- BMI 수치를 정수로 버림으로써 경계값 처리
- 미사용 쿼리 삭제

+++ project/back/health/src/main/java/com/ssafy/health/domain/body/BodyType/repository/BodyTypeRepository.java
@@ -14,5 +14,5 @@ public interface BodyTypeRepository extends JpaRepository<BodyType, Long> {
             WHERE b.gender = :gender AND :bmi BETWEEN b.minFatRatio AND b.maxFatRatio
             ORDER BY b.id ASC LIMIT 1
             """)
-    Optional<BodyType> findByBmiAndGender(Float bmi, Gender gender);
+    Optional<BodyType> findByBmiAndGender(Integer bmi, Gender gender);
 }
